feat: Improve the point's and ranking system.

+++ database/migrations/2025_02_28_225532_add_rank_to_form_data_table.php
@@ -12,11 +12,21 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('form_data', function (Blueprint $table) {
-            $table->integer('rank')->nullable()->default(0)->after('status');
+            if (!Schema::hasColumn('form_data', 'rank')) {
+                $table->integer('rank')->nullable()->default(0)->after('status');
+            }
         });
 
         Schema::table('form_field_groups', function (Blueprint $table) {
-            $table->boolean('requires_auth')->default(false)->after('authenticator');
+            if (!Schema::hasColumn('form_field_groups', 'requires_auth')) {
+                $table->boolean('requires_auth')->default(false)->after('authenticator');
+            }
+        });
+
+        Schema::table('form_fields', function (Blueprint $table) {
+            if (!Schema::hasColumn('form_fields', 'points')) {
+                $table->integer('points')->nullable()->default(0)->after('max');
+            }
         });
     }
 
@@ -26,11 +36,21 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('form_data', function (Blueprint $table) {
-            $table->dropColumn(['rank']);
+            if (Schema::hasColumn('form_data', 'rank')) {
+                $table->dropColumn(['rank']);
+            }
         });
 
         Schema::table('form_field_groups', function (Blueprint $table) {
-            $table->dropColumn(['requires_auth']);
+            if (Schema::hasColumn('form_field_groups', 'requires_auth')) {
+                $table->dropColumn(['requires_auth']);
+            }
+        });
+
+        Schema::table('form_fields', function (Blueprint $table) {
+            if (Schema::hasColumn('form_fields', 'points')) {
+                $table->dropColumn(['points']);
+            }
         });
     }
 };
